+++ browser/main.js
@@ -1,4 +1,5 @@
 var insertCss = require('insert-css')
+var merge = require('lodash/merge')
 var ipc = require('electron').ipcRenderer
 var remote = require('electron').remote
 var dialog = require('electron').dialog
@@ -16,7 +17,35 @@ window.addEventListener('hashchange', function (ev) {
   localStorage.setItem('location', location.hash)
 })
 
-document.addEventListener("DOMContentLoaded",  function () {
+var $overlay = document.getElementById('overlay')
+var $welcome = document.getElementById('welcome')
+var $map = document.getElementById('container')
+
+var showedWelcome = localStorage.getItem('showedWelcome')
+if (!showedWelcome) {
+  localStorage.setItem('showedWelcome', true)
+  welcomeScreen($overlay, $welcome, $map)
+} else openMap()
+
+updateSettings()
+
+var serverUrl = 'http://' + remote.getGlobal('osmServerHost')
+var id = iD.Context()
+  .assetPath('vendor/iD/')
+  .preauth({url: serverUrl})
+  .minEditableZoom(14)
+
+var customDefs = id.container()
+  .append('svg')
+  .style('position', 'absolute')
+  .style('width', '0px')
+  .style('height', '0px')
+  .attr('id', 'custom-defs')
+  .append('defs')
+
+customDefs.append('svg')
+
+id.ui()(document.getElementById('container'), function onLoad () {
   if (document.querySelector("a[href*='iD/issues']")) {
     document.querySelector("a[href*='iD/issues']")
       .setAttribute('href', 'https://github.com/digidem/mapeo-desktop/issues')
@@ -33,51 +62,24 @@ document.addEventListener("DOMContentLoaded",  function () {
     document.querySelector(".overlay-layer-attribution a").
       innerHTML = i18n('feedback-contribute-button')
   }
+  var aboutList = id.container().select('#about-list')
+  // Update label on map move
+  var map = id.map();
+
+  var latlon = aboutList.append('li')
+  .append('span')
+  .text(latlonToPosString(map.center()))
+
+  map.on('move', function () {
+    var pos = map.center()
+    var s = latlonToPosString(pos)
+    console.log(s)
+    latlon.text(s)
+  })
+  updateSettings()
 })
 
-var parser = new DOMParser()
-
-var serverUrl = 'http://' + remote.getGlobal('osmServerHost')
-
-var id = iD.Context()
-  .assetPath('vendor/iD/')
-  .preauth({url: serverUrl})
-  .minEditableZoom(14)
-
-// window.locale.en.inspector.view_on_osm = 'View element source XML'
-
-var $overlay = document.getElementById('overlay')
-var $welcome = document.getElementById('welcome')
-var $map = document.getElementById('container')
-
-var showedWelcome = localStorage.getItem('showedWelcome')
-if (!showedWelcome) {
-  localStorage.setItem('showedWelcome', true)
-  welcomeScreen($overlay, $welcome, $map)
-} else openMap()
-
-id.ui()(document.getElementById('container'))
-
-function myOnBeforeLoad () {
-  id.save()
-}
-
-var parser = new DOMParser()
-var customDefs = id.container()
-  .append('svg')
-  .style('position', 'absolute')
-  .style('width', '0px')
-  .style('height', '0px')
-  .attr('id', 'custom-defs')
-  .append('defs')
-
-customDefs.append('svg')
-updateSettings()
-
-
-function myOnBeforeLoad () {
-  id.save()
-}
+window.onbeforeunload = function () { id.save() }
 
 function openMap () {
   $overlay.innerHTML = overlay()
@@ -101,12 +103,22 @@ function updateSettings () {
   var imagery = ipc.sendSync('get-user-data', 'imagery')
   var icons = ipc.sendSync('get-user-data', 'icons')
 
-  if (presets) iD.data.presets = presets
+  if (presets) {
+    if (!id) iD.data.presets = merge(iD.data.presets, presets)
+  }
   if (customCss) insertCss(customCss)
-  if (imagery) iD.data.imagery = imagery
+  if (imagery) {
+    if (!imagery.id) {
+      imagery = imagery.map(function (i) { i.id = i.name; return i; })
+      console.warn('imagery.json should include an id -- using name instead.', imagery)
+    }
+    iD.data.imagery = imagery
+  }
   if (icons) {
+    var parser = new DOMParser()
     var iconsSvg = parser.parseFromString(icons, 'image/svg+xml').documentElement
-    customDefs.node().replaceChild(iconsSvg, customDefs.node().firstChild)
+    var defs = customDefs && customDefs.node()
+    if (defs) defs.replaceChild(iconsSvg, defs.firstChild)
   }
   log('settings updated')
 }
@@ -130,4 +142,12 @@ function zoomToLatLonResponse (_, lat, lon) {
   setTimeout(function () {
     id.map().zoom(15)
   }, 1000)
+}
+
+function latlonToPosString (pos) {
+  pos[0] = (Math.floor(pos[0] * 1000000) / 1000000).toString()
+  pos[1] = (Math.floor(pos[1] * 1000000) / 1000000).toString()
+  while (pos[0].length < 10) pos[0] += '0'
+  while (pos[1].length < 10) pos[1] += '0'
+  return pos.toString()
 }
