--- id_monkey_patches/locales/es.json
@@ -1,15 +1,0 @@
-{
-  "commit": {
-    "upload_explanation": "Una vez guardado, puedes sincronizar tus cambios con otros.",
-    "upload_explanation_with_user": "Una vez guardado, puedes sincronizar tus cambios con otros."
-  },
-  "inspector": {
-    "view_on_osm": "Ver fuente XML del elemento"
-  },
-  "success": {
-    "edited_osm": "¡Mapeo editado!",
-    "just_edited": "¡Acaba de editar Mapeo!",
-    "view_on_osm": "Ver en OSM",
-    "help_html": "Puedes sincronizar tus cambios con otros ahora con el botón \"Sincronizar\"."
-  }
-}
