+++ lib/locale.js
@@ -12,6 +12,7 @@ module.exports =  {
 function load () {
   var english = read('en')
   var locale = app.getLocale()
+  locale = locale.split('-')[0]
   var messages
   try {
     var translations = read(locale)
