+++ package.json
@@ -63,7 +63,7 @@
     "hyperlog": "^4.12.0",
     "hyperlog-index": "^5.2.2",
     "hyperlog-sneakernet-replicator": "^1.1.2",
-    "id-mapeo": "^1.0.2",
+    "id-mapeo": "^1.0.3",
     "insert-css": "^1.0.0",
     "level": "^1.4.0",
     "levelup": "^1.3.1",
@@ -97,7 +97,6 @@
     "appdmg": "0.4.5"
   },
   "devDependencies": {
-    "@gmaclennan/concat": "^1.0.0",
     "cross-env": "^1.0.8",
     "cross-spawn": "^4.0.2",
     "devtron": "^1.2.0",
